Fix minor issues.

+++ src/command/info/info.c
@@ -106,7 +106,7 @@ Data types and structures
 typedef struct InfoRepoData
 {
     unsigned int key;                                               // User-defined repo key
-    CipherType cipher;                                              // Encryption type (0 = none)
+    CipherType cipher;                                              // Encryption type
     const String *cipherPass;                                       // Passphrase if the repo is encrypted (else NULL)
     int stanzaStatus;                                               // Status code of the the stanza on this repo
     unsigned int backupIdx;                                         // Index of the next backup that may be a candidate for sorting
@@ -640,7 +640,7 @@ stanzaInfoList(List *stanzaRepoList, const String *backupLabel, unsigned int rep
         VariantList *repoSection = varLstNew();
 
         int stanzaStatusCode = -1;
-        unsigned int stanzaCipherType = 0;
+        uint64_t stanzaCipherType = 0;
 
         // Set the stanza name and initialize the overall stanza variables
         kvPut(varKv(stanzaInfo), KEY_NAME_VAR, VARSTR(stanzaData->name));
